update: DB処理をモジュール化
